+++ NEWS.md
@@ -343,8 +343,7 @@ The following errors are caused by breaking changes.
   and `names_to` is `NULL`, the names define row names. If `names_to`
   is supplied, they are assigned in the column name as before.
 
-* `vec_cbind()` now binds row names if they are congruent across
-  inputs. If the row names are not identical that's an error.
+* `vec_cbind()` now uses the row names of the first named input.
 
 * The `c()` method for `vctrs_vctr` now throws an error when
   `recursive` or `use.names` is supplied (#791).